Fixed windows compilation issues.  Also fixed all the warnings that svs has on windows....

+++ src/soar_interface.h
@@ -157,7 +157,7 @@ inline bool soar_interface::get_val(Symbol *sym, float &v) {
 		return true;
 	}
 	if (is_int(sym)) {
-		v = sym->ic.value;
+		v = float(sym->ic.value);
 		return true;
 	}
 	return false;
@@ -169,7 +169,7 @@ inline bool soar_interface::get_val(Symbol *sym, double &v) {
 		return true;
 	}
 	if (is_int(sym)) {
-		v = sym->ic.value;
+		v = double(sym->ic.value);
 		return true;
 	}
 	return false;
